requirements.txt
Django==2.0.1
django-dotenv==1.4.2
django-timezone-field==2.0
pkg-resources==0.0.0
pytz==2017.3
